refactor(core): keystone — invert DreamCycle to CycleCtx + relocate snapshot/cycle/search DTOs to types/ (#816)
Wave 2 (#816) S1, in-monolith keystone (no crates carved yet — E.3–E.5 follow
in this slice). Breaks every would-be cross-crate cycle while the code is still
one crate, so the carve and the cycle-break are independently verifiable.

- DreamCycle inversion: `DreamCycle::run` now takes `&dyn CycleCtx` (a new
  5-method me-traits-bound read surface: time_window / prior_wisdom /
  prior_reports / list_undreamt_in_period / outcome_counts_batch, derived
  mechanically from the two shipped impls). `CycleContext` implements it. This
  is a deliberate, gated public-API break (see #816 plan Correction 4 / XR-3),
  flattening the `ctx.dream().*` indirection. Without it, `me-traits` would
  depend on `me-consolidate` once the crates split.
- DTO relocation to `types/` (the me-types-destined home):
  * snapshot: the 8 serde DTOs → `types/snapshot.rs`; the rusqlite+fs IO stays
    in `engine/snapshot.rs` (→ me-backend-sqlite in S2).
  * cycle report: `engine/cycle/report.rs` → `types/cycle_report.rs` (git mv,
    blame preserved); `engine::cycle` re-exports it for the public API.
  * search-result vocabulary (SearchResult/SearchQuery/QueryResponse/
    QueryDiagnostics/SearchMode/MatchType/RRF_K) → `types/search.rs`; the
    fusion logic (`rrf_merge`, hybrid search) stays in `search/`.
- This breaks the one real stray edge `traits -> search::hybrid::SearchResult`:
  `traits.rs` and `types/` now have zero `crate::engine::`/`crate::search::`
  code imports.

Verified: `cargo test -p memory-engine --all-features` = 1568 passed, 0 failed
(== pre-change baseline, no test-count regression); `cargo fmt --check` clean.
