+++ memory-engine/lib/memory-engine/src/engine/cycle/mod.rs
@@ -1,6 +1,7 @@
 //! The dream-cycle subsystem (Phase 5a).
 //!
-//! - [`report`] — the delta-based [`CycleReport`] vocabulary (R7).
+//! - the delta-based [`CycleReport`] vocabulary (R7) now lives in
+//!   [`crate::types::cycle_report`] (Wave 2 #816).
 //!
 //! Further submodules (`context`, `apply`, `dbscan`, `default_impl`) are added by
 //! later tasks. Public items are re-exported flat from the crate root (`lib.rs`).
@@ -10,12 +11,11 @@ mod context;
 mod dbscan;
 mod default_impl;
 mod llm_impl;
-mod report;
 
-pub use context::CycleContext;
-pub use default_impl::DefaultDreamCycle;
-pub use llm_impl::LlmDreamCycle;
-pub use report::{
+pub use crate::types::cycle_report::{
     ApplyResult, CycleAnomaly, CycleDelta, CycleMetadata, CycleOutcome, CycleReport,
     IMPORTANCE_STEP, IdentityOutput, MAX_ADJUSTMENT, SkipReason, TimeWindow,
 };
+pub use context::CycleContext;
+pub use default_impl::DefaultDreamCycle;
+pub use llm_impl::LlmDreamCycle;
